Refactor and add comments
