+++ app/src/main/java/com/benallouch/yunar/db/ArticleDao.kt
@@ -1,6 +1,8 @@
 package com.benallouch.yunar.db
 
 import androidx.room.*
+import com.benallouch.yunar.db.entity.ArticleDbEntity
+import com.benallouch.yunar.db.entity.TotalResultsEntity
 
 @Dao
 interface ArticleDao {
