+++ app/src/main/java/com/benallouch/yunar/db/ArticleRoomDatabase.kt
@@ -5,6 +5,9 @@ import androidx.room.Database
 import androidx.room.Room
 import androidx.room.RoomDatabase
 import androidx.room.TypeConverters
+import com.benallouch.yunar.db.entity.ArticleDbEntity
+import com.benallouch.yunar.db.entity.NewsSourceDbEntity
+import com.benallouch.yunar.db.entity.TotalResultsEntity
 
 @Database(entities = [ArticleDbEntity::class, NewsSourceDbEntity::class, TotalResultsEntity::class], version = 1)
 @TypeConverters(DateConverter::class)
