+++ app/src/main/java/com/benallouch/yunar/db/DateConverter.kt
@@ -3,6 +3,9 @@ package com.benallouch.yunar.db
 import androidx.room.TypeConverter
 import java.util.*
 
+/**
+ * Mandatory to persist date objects in [ArticleRoomDatabase]
+ */
 class DateConverter {
     @TypeConverter
     fun toDate(dateLong: Long?): Date? {
