+++ app/src/main/java/com/benallouch/yunar/db/entity/ArticleDbEntity.kt
@@ -1,4 +1,4 @@
-package com.benallouch.yunar.db
+package com.benallouch.yunar.db.entity
 
 import androidx.room.Embedded
 import androidx.room.Entity
@@ -11,11 +11,11 @@ data class ArticleDbEntity(
         @Embedded
         var source: NewsSourceDbEntity,
         var author: String?,
-        var title: String,
-        var description: String,
-        var url: String,
+        var title: String?,
+        var description: String?,
+        var url: String?,
         var urlToImage: String?,
-        var publishedAt: Date,
+        var publishedAt: Date?,
         var content: String?
 )
