+++ app/src/main/java/com/benallouch/yunar/ui/main/adapter/RecyclerViewPager.kt
@@ -3,10 +3,14 @@ package com.benallouch.yunar.ui.main.adapter
 import androidx.recyclerview.widget.GridLayoutManager
 import androidx.recyclerview.widget.RecyclerView
 
+/**
+ * @param loadMore lambda function to trigger a new api call with the next page
+ * @param recyclerView our Recyclerview to which we attach the pagination
+ */
 class RecyclerViewPager(
         private val recyclerView: RecyclerView,
         private val loadMore: (Int) -> Unit
-        ) : RecyclerView.OnScrollListener() {
+) : RecyclerView.OnScrollListener() {
 
     var isLoading = false
     var lastPage: Int = 1
@@ -32,6 +36,7 @@ class RecyclerViewPager(
             if (isLoading) return
 
             if (lastVisibleItemPosition + 4 >= allItemsSize) return
+
             if (!isLoading && ((lastVisibleItemPosition + 4) >= totalItemCountInLayout)) {
                 isLoading = true
                 ++lastPage
