+++ data/src/main/java/com/benallouch/data/repository/ArticlesRepository.kt
@@ -21,7 +21,9 @@ class ArticlesRepository(
             shouldFetchRemotely -> {
                 newsResponse = remoteDataSource.getArticles(apiKey, page)
                 //The APi doesn't return ids so we generate some ids so we can use them in Diffutil and DB persistence
-                newsResponse.articles.forEach { it.articleId = generateId(it) }
+                newsResponse.articles.forEach {
+                    it.articleId = generateId(it) ?: UUID.randomUUID().toString()
+                }
                 localDataSource.saveArticles(newsResponse.articles)
                 localDataSource.saveTotalItems(newsResponse.totalResults)
             }
@@ -36,8 +38,8 @@ class ArticlesRepository(
 
     }
 
-    private fun generateId(article: Article): String {
-        return article.description.hashCode().toString() + article.publishedAt.hashCode().toString()
+    private fun generateId(article: Article): String? {
+        return article.description?.hashCode().toString() + article.publishedAt?.hashCode().toString()
     }
 
 }
