--- usecase/src/main/java/com/benallouch/usecase/MyClass.kt
@@ -1,4 +1,0 @@
-package com.benallouch.usecase
-
-public class MyClass {
-}
